complete access level and profile view and control session

+++ api/view_user.py
@@ -1,24 +1,20 @@
-import random
 from accounts import otp_check
 from accounts.models import User
 from random import randint
-from django.conf import settings
-from django.contrib import messages
-from datetime import date, datetime, tzinfo
+from datetime import date
 from movie.models import SessionUser
-from django.shortcuts import render, redirect
+from django.shortcuts import redirect
 from django.contrib.sessions.models import Session
 from django.utils.translation import gettext_lazy as _
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth import login, logout, authenticate
-from .seializer_user import OtpSerializer,PhoneSerializer, UserSerializer,PaidSerializer,SessionSerializer
+from django.contrib.auth import login, logout
+from .seializer_user import OtpSerializer,PhoneSerializer, UserSerializer,PaidSerializer,SessionSerializer, SaveSerializer
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
-from rest_framework import status, views
+from rest_framework import status
 from rest_framework.authtoken.models import Token
 
 from accounts.models import User
-from movie.models import Serial, Movie, Save
+from movie.models import Save
 from django.contrib.sessions.models import Session
 from django.shortcuts import get_object_or_404
 
@@ -69,6 +65,7 @@ def sms(request):
     try:
         phone = request.session.get('phone_number')
         user = User.objects.get(phone_number = phone)
+        token = Token.objects.get_or_create(user=user)
 
         if request.method == 'POST':
             form = OtpSerializer(data=request.data)
@@ -83,7 +80,6 @@ def sms(request):
                     user.save()
                     del request.session['phone_number']
                     login(request, user)
-                    token = Token.objects.get_or_create(user=user)
                     SessionUser.objects.get_or_create(
                                         user = user,
                                         session_key =Session.objects.get(session_key = request.session.session_key),
@@ -91,7 +87,6 @@ def sms(request):
                                         os = f'{request.user_agent.os.family}-{request.user_agent.os.version_string}',
                                         date_joiin = date.today(),
                                         ip_device =request.META['REMOTE_ADDR'],
-                                        token = token
                                     )
 
                     return redirect('api:index')
@@ -114,6 +109,9 @@ def sms(request):
 
 @api_view(['GET'],)
 def signout(request):
+    if not request.user.is_authenticated:
+        return redirect('api:phone')
+        
     logout(request)
     return redirect('api:phone')
 
@@ -121,6 +119,9 @@ def signout(request):
 
 @api_view(['GET'],)
 def profile(request):
+    if not request.user.is_authenticated:
+        return redirect('api:index')
+
     pk = request.user.pk
 
     user = get_object_or_404(User, pk=pk)
@@ -134,23 +135,27 @@ def profile(request):
     
     
     saves = Save.objects.filter(user = request.user)
+    save = SaveSerializer(saves, many=True).data
 
-    srz = (form, paid,sessn)
+    srz = (form, paid,sessn,save)
     return Response(srz, status=status.HTTP_200_OK)
 
 
 
 
 
-
-def remove_session(request):
+@api_view(['POST'],)
+def delete_session(request):
     if request.method == 'POST':
-        key = request.POST.get('key')
+        key = request.data('key')
         try:
-            session = Session.objects.get(session_key = key)
-            session.delete()
-            return redirect('movie:user')
+            sessn = SessionUser.objects.get(user = request.user, session_key__session_key=key)
+            if sessn:
+                session = Session.objects.get(session_key = key)
+                session.delete()
+                return redirect('api:profile')
+            return redirect('api:profile')
         except:
-            return redirect('movie:user')
+            return redirect('api:profile')
     else:           
-        return redirect('movie:user')
+        return redirect('api:profile')
